requirements.txt
mysql-connector-python
redis
pymongo
